Search widget integration: quick fix, cast std::distance to int
As suggested by the buddy. Android Studio did not complain, but the iOS
compiler does.

Buddy: Andrew Gurney

+++ src/SearchMenu/View/SearchServices.cpp
@@ -71,7 +71,7 @@ namespace ExampleApp
 					HandleSpecialCases(*it, desc, icon);
 
 					m_searchResults.push_back(SearchServicesResult(name, desc, icon,
-																   std::distance(sdkResults.begin(), it),
+																   static_cast<int>(std::distance(sdkResults.begin(), it)),
 																   *it));
 				}
